Fixed a few more MSVS warnings; Down to 7 warnings that are not easy to fix

+++ src/inlines.h
@@ -372,11 +372,11 @@ inline bool line_poly_intersect(vector3d const &v1, point const &p1, point const
 }
 
 inline bool point_in_ellipse(point const &p, point const &center, float rx, float ry) {
-	double const dx(center.x - p.x), dy(center.y - p.y);
+	double const dx((double)center.x - p.x), dy((double)center.y - p.y);
 	return (dx*dx/((double)rx*rx) + dy*dy/((double)ry*ry) <= 1.0);
 }
 inline bool point_in_ellipse_risq(point const &p, point const &center, float rx_inv_sq, float ry_inv_sq) {
-	double const dx(center.x - p.x), dy(center.y - p.y);
+	double const dx((double)center.x - p.x), dy((double)center.y - p.y);
 	return (dx*dx*(double)rx_inv_sq + dy*dy*(double)ry_inv_sq <= 1.0);
 }
